Update day07.clj
I was going through your Day07 solution to help learn how you solved it. 

I noticed that when I ran the code it didn't produce the correct solution. The change submitted here fixes the problem.

With my data set I receive this when I first ran your code:
;; Order of step QPFKWJSVUXEMNIHGTYDOZACRLB

With the above change I get the following which is correct
;; The correct answer is PFKQWJSVUXEMNIHGTYDOZACRLB